+++ multi_vendor.go
@@ -11,6 +11,7 @@ var vendors = []*vendor{newJuniperJUNOS(), newCiscoIOSXR()}
 type vendor struct {
 	name               string
 	loginCheckRequired bool
+	sendLoginCheck     func(*JCtx, *grpc.ClientConn) error
 	dialExt            func(*JCtx) grpc.DialOption
 	subscribe          func(*grpc.ClientConn, *JCtx, chan<- bool) SubErrorCode
 }
@@ -33,6 +34,7 @@ func newJuniperJUNOS() *vendor {
 	return &vendor{
 		name:               "juniper-junos",
 		loginCheckRequired: true,
+		sendLoginCheck:     loginCheckJunos,
 		dialExt:            nil,
 		subscribe:          subscribeJuniperJUNOS,
 	}
@@ -42,6 +44,7 @@ func newCiscoIOSXR() *vendor {
 	return &vendor{
 		name:               "cisco-iosxr",
 		loginCheckRequired: false,
+		sendLoginCheck:     nil,
 		dialExt:            getXRDialExtension,
 		subscribe:          subscribeCiscoIOSXR,
 	}
